requirements.txt
streamlit
langgraph
opencv-python
ultralytics
deepface
geohash2
pandas
numpy
watchdog
google-generativeai>=0.3.0
reportlab>=4.0.0
websockets>=12.0
python-dotenv>=1.0.0

